Modularize cloud core boundaries
