+++ apps/desktop/src/main/cloud/http-routes/types.ts
@@ -17,10 +17,12 @@ export type CloudApiRouteTools = {
   readJsonBody(req: IncomingMessage, maxBodyBytes: number): Promise<Record<string, unknown>>
   readString(value: unknown): string | null
   readRecord(value: unknown): Record<string, unknown> | null
+  readStringArray(value: unknown): string[] | null
   readOptionalDate(value: unknown): Date | null
   readApiTokenScopes(value: unknown): ApiTokenScope[] | null
   readOptionalCloudProjectSource(body: Record<string, unknown>): CloudProjectSourceInput | null | undefined
   parseLimit(url: URL): number | undefined
+  parseTagIds(url: URL): string[]
   writeJson(res: ServerResponse, status: number, body: unknown, origin?: string | null): void
   writeError(res: ServerResponse, status: number, message: string, origin?: string | null): void
   writePolicyError(res: ServerResponse, status: number, message: string, policyCode: string, origin?: string | null): void
