+++ apps/desktop/src/main/cloud/postgres-domains/schema.ts
@@ -1,0 +1,9 @@
+import type { SchemaMigrationRecord } from '../control-plane-store.ts'
+import { iso, type QueryRow } from './shared.ts'
+
+export function migrationFromRow(row: QueryRow): SchemaMigrationRecord {
+  return {
+    id: String(row.id),
+    appliedAt: iso(row.applied_at),
+  }
+}
